in setting in usb port access add -> enable/disable option in a single switch if it enable it should say enable if its disabled it should say disable .
for malware detection give options like scan and auto scan if malware is detected automatically pause transfrer and temporaryly block the code untill usb is removed.

for authentication control give options like enable and disable.

for alerts

give real time notifications with pop ups
for usb is connected or disconnected
for malware is detected or file is safe to use
for monitoring give a real time popup which shows how much data is transfered whats the total space in drive that is connected and other usb/drive status
such as max speed for the transfer which is recorded.